Add files via upload
This is the lead preset used to create the track "Korridor - Vacuum Decay" released on Hypnus Records in 2014. 

The preset is opened with the VST synth "Vember Audio Surge" I used version 1.5.2 when i made this preset so work from that version and upwards if you want to open it.

Nowadays "Vember Audio Surge" is open source and can be found on this link: https://surge-synthesizer.github.io/

The preset folder where all the presets are stored is sometimes hidden in the Virtualstore folder in AppData. For me the address looks like this.

C:\Users\Fabian Kempe\AppData\Local\VirtualStore\Program Files (x86)\Steinberg\VstPlugins\Vember Audio\Surge 1.5.2\surgedata\patches_user

Meanwhile there is a \surgedata\patches_user folder also located in the install folder for surge, but presets arent saved there when you work with the synth. Instead they are saved to the address above. Keep this in mind of you want to add new presets.

//Fabian aka Korridor